Added comments and SHUTDOWN_MAX_WAIT
Added a couple comments to help me, and future editors, to understand the results of checking if a server screen exists or not.
Added a new variable per server. SHUTDOWN_MAX_WAIT. This allows the admin to customize how long until the countdown exists with an error. If the server screen clears/closes before the countdown reaches 0, the countdown discontinues and proceeds to the next step.
Added variable for each server.
To-Do: Implement another input to allow the server admin to forcibly kill a screen.