Manual Server Overrides are in!
Got servers you want to auto start when the computer/system starts, but not all of them? Now you can!
On each server in the servers.cfg file, you have the option in SERVER_ENABLED of "true", "false", and now "manual".
When the server is set as manual, it will not auto start unless told so by command with OVERRIDE at the end.
Example: service mcservers server start 2 OVERRIDE.
If the variable is set to "true" or "false", OVERRIDE is ignored and treated as normal. If "false" server will never start until servers.cfg is changed to have the option set to "true" or auto start, or "manual" commanded to start with OVERRIDE variable.