Minor fixes
Fixed some minor error errors here and there.
In the last update, I didn't realize the $VARIABLEs in the servers.cfg file was not being created. Added escape codes for these to spawn correctly.
Fixed an error in the notification of where the servers.cfg file is created. Though it has been being created, the final notification had a couple variable typos.